Problem. Solved.
Changed the baseline URL suffix value to the result I got after ~400
iterations. Let it run again, watched the results, and noticed a blank
around iteration #66...found my answer.

It's terribly slow, and I want to revisit the general structure of it at
a later date. It was a great learning experience, though.